requirements.txt
asgiref==3.2.7
colorama==0.4.3
dj-database-url==0.5.0
Django==3.0.6
django-cors-headers==3.3.0
django-heroku==0.3.1
djangorestframework==3.11.0
gunicorn==20.0.4
mccabe==0.6.1
psycopg2==2.8.5
pytz==2020.1
sqlparse==0.3.1
toml==0.10.1
vue==0.0.1
whitenoise==5.1.0
